Correctly raise/clear alerts with purppura.
Previously we calculated an alert-ID which was "unique" based upon
the input line.

So we'd generate an ID by hashing:

        localhost must run rsync

However this test would be expanded into two:

        127.0.0.1 must run rsync
        fe80::1 must run rsync

This meant that the second alert would overwrite the first, as the
hashed IP would be identical.

Update to use the target we _actually_ executed against too.