Reworked the parser to ensure it is globally used.
This commit changes the interface of the parser, such that there
are now two public methods:

* ParseFile(filename)
* ParseLine(line)

The callers have all been updated to use this new input, which
means that the redis-based watcher no longer tries to parse input
lines itself - the parser does it, as it should.

This solves the problem of invalid lines killing the worker because
it doesn't handle them the same way as the parser.

This closes #7.