Unified (optional) arguments for protocol-tests.
Almost every test we had allowed the use of an optional argument
to alter its behaviour.  We've now unified how this works for
all tests, except the DNS one, to use a common form:

    target must run xxx WITH $NAME $OPTION WITH $NAME2 $OPTION2

This results in no change for the http-probe, as we always allowed:

* with status 123
* with content "something here"

For the other tests it mostly means that this:

     on port 1234

becomes

     with port 1234

The TLS testing used to be disabled for IMAPS + POP3S via:

     must run pop3s insecure

But now becomes:

     with tls insecure

A little more verbose, but more standard.

Added test-cases for the parser.