Unify the DNS test.
Previously the DNS-test was the odd one out, requiring it to be
written in a non-standard way.  Now that we've unified all the
optional arguments to be of the form:

  with $OPTION_NAME $OPTION_VALUE

We can use the same approach in the DNS test, to allow:

       NS must run dns with lookup steve.fi with type A with result '176.9.183.100'

This is more verbose, but hopefully the consistency will make up
for it.