Unified parser and argument passing.
We've previously written a general purpose handler for extracting
(optional) arguments passed to tests, such that:

        foo must run bar with $option_name $option_value

However each test had to parse the optional arguments separately,
and it made more sense to do it once in the parser.

We've now reworked the protocol-tester API to allow passing the
(parsed) test, arguments and all, to the test-job.  This simplifies
a lot of the code.

The only complication is that both the protocol-handlers and the
parser needed to refer to the test-structure object, and that
introduced a cyclic dependency issue.  This has been resolved by
putting the test-structure into a new package.

This closes #8.