Unified HTTP-test with all others.
We no longer perform the DNS-resolution in our HTTP-test, instead
it is processed exactly like the others:

* The `run_test` method in `core.go` does DNS resolution.
* For each result we run the HTTP-test.

This has been achieved by no longer treating `http://xxx` as a
special case - instead if the test-runner sees something that has
a "://" string in it the target is treated as an URI and the hostname
is extracted & used.

This closes #14.