Allow persistent notifiers.
This commit updates the API of the notifier class to allow
a Setup() method to be called at construction-time.  The intention
is that this can be used to join chat-servers, etc, and that
open connection reused to actually trigger notifications.

The IRC & MQ notifiers have been updated to work in this way,
which closes #17.