Allow each protocol-handler to return the arguments it supports.
We've updated the signature of the protocol-handler-object to
add a new method:

* `Arguments() []string`

This returns an array of the arguments which the handler supports,
and is designed to allow the parser to throw an error if an input
test tries to use an unsupported argument.

i.e. This is a robustness __and__ documentation-change.

This closes #18.