Allow removing passwords prior to notification.
This commit adds a new method to the Test.test type, to allow
the test to be retrieved as a string.  The point of this is to
remove any password from the input, to allow it to be posted
to an MQ-queu, purppura notifier, or similar.

While people writing the tests need to know passwords it is
possible that the sysadmins/on-call people who react to them
do not.

It seems safer to censor them regardless.

This closes #22.