Update our MQ-handling to ensure messages are published.
In the past, simplifying, we would:

* Run a test
* Publish the result to MQ
* Terminate

If the termination of our process happened "too quickly" the message
we had published would have been lost/dropped - because it would have
been processed asyncronously.

Rather than allow that we've now updated our MQ-handling such that
each publish actually awaits delivery of the message, via a busy-loop.

This closes #27, albeit in a horrible way.  I've updated the
appropriate issue in the MQ-library to mention this "solution"
too:

https://github.com/yosssi/gmq/issues/11