Sleep between failing tests.
Prior to this commit a failing test would be retried several times
(5 by default), such that a transient failure would not raise an
alert.

However if a service was restarting (specifically apache2 restarting
on my servers) this would result in retries immediately.  Instead
we need to sleep a little between failing tests.

This means we go from:

  1/5: Connection refused
  2/5: Connection refused
  3/5: Connection refused
  4/5: Connection refused
  5/5: Connection refused

To:

   1/5 connection-refused
   [sleep]
   2/5 - OK (because apache has restarted in the few seconds of delay).

I've also moved some flags from the test-options structure, because
tests don't need to know about them.