Updated to drop the notifier/ package.
In the past we allowed pluggable notifiers, but now we don't.  So
it doesn't make sense to keep the extra package around when the
core of that package is a single function.

So the notifier-code has been moved into the worker where it is
more logical to find it.

This closes #31.