Updated our SMTP-probe to perform SMTP authentication.
This commit updates the SMTP-probe to support optional
SMTP-authentication.  If a username/password are supplied
they will be tested against the remote host.

NOTE: Authentication requires the use of STARTTLS, to avoid
submitting credentials in plaintext, so there is an optional
`with tls insecure` support to disable any warnings about
unknown CAs, or hostname mismatches.

This closes #33.